% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression function}
\usage{
ridge(X, y, lambdas = 0)
}
\arguments{
\item{form, }{an R  formula to construct the regression}

\item{data, }{the data to use for the regression}

\item{lambda, }{an optional set of lambdas to use for regression}
}
\description{
Assignment 2 for BIS 557
}
\details{
Goal: compute a matrix of regression values for a numeric data set (X) using ridge regression
}
